% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca-r2.R
\name{lod.root2}
\alias{lod.root2}
\title{Single pollutant \code{sqrt(2)} imputation.}
\usage{
lod.root2(formula, data, exposure, lod, type)
}
\arguments{
\item{formula}{A R formula in the form outcome ~ exposure + covariates.}

\item{data}{A data.frame that contains the variables \code{formula}
references.}

\item{exposure}{Name of the exposure variable}

\item{lod}{Name of the lod variable}

\item{type}{The type of regression to perform. Acceptable options are
"linear" and "logistic".}
}
\description{
lod.ro2 is a helper function that does \code{sqrt(2)} single imputation for
single polluatant models. It is pretty common to recode rehistered values
below the lod as \code{lod / sqrt(2)}. The function's primary purpose is to
compare it to clmi.
}
