% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmi.R
\name{pool.clmi}
\alias{pool.clmi}
\title{Calculate pooled estimates from \code{clmi.out} objects}
\usage{
pool.clmi(clmi.out, regression.type = c("logistic, linear"),
  outcome.covars = NULL)
}
\arguments{
\item{clmi.out}{an object generated by clmi.}

\item{regression.type}{type of regression to pool. Valid types are
"logistic" and "linear".}

\item{outcome.covars}{a character vector of variables associated
with the outcome. Only include variable names for variables that were
not used during imputation (variables that were used for imputation
are automatically adjusted for).}
}
\description{
Calculate pooled estimates from \code{clmi.out} objects
}
\examples{
# continue example from clmi
# fit model on imputed data and pool results
library(lodmi)
data("toy_data")
covars <- c("smoking", "gender")
clmi.out <- clmi(toy_data, "poll", "batch1", "case_cntrl", covars,
                   lod.info = data.frame(batch.info = c("1","0"),
                     lod = c(0.8, 0.65)), 20, 12345,
                   t.function = function(x) log(x))
results <- pool.clmi(clmi.out, "logistic", NULL)

results$output
}
