% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy.data}
\alias{toy.data}
\title{Synthetic toy data for clmi}
\format{A data.frame with 100 obseration on 6 variables:
\describe{
  \item{id}{Patient ID number.}
  \item{case_cntrl}{Patient's case-control status. Either 1 or 0.}
  \item{poll}{Concentration of pollutant in patient's blood sample.}
  \item{smoking}{Smoking status. Either 1 or 0.}
  \item{gender}{Gender. 1 for male, 0 for female.}
  \item{batch1}{Batch status. Integer}
}}
\usage{
toy.data
}
\description{
Synthetic toy data for clmi
}
\keyword{datasets}
