% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clmi.R
\name{clmi}
\alias{clmi}
\title{Censored Likelihood Multiple Imputation}
\usage{
clmi(formula, df, outcome, exposure, lod, seed, n.imps = 5,
  t.function = function(x) x)
}
\arguments{
\item{df}{a data.frame with exposure concentration, batch number,
covariates used in imputation, precision variables.}

\item{outcome}{name of the outcome variable in df.}

\item{exposure}{name of the exposure variable in df.}

\item{seed}{for reproducability}

\item{n.imps}{number of datasets to impute. Default is 5.}

\item{t.function}{transformation function to apply to the exposure to make
it normally distributed. Default is no transformation}

\item{contam.covars}{covariates to use in imputing exposure. Can be NULL.}

\item{lod.info}{a two column data.frame. The first column must be named
"batch.info" corresponds to each batch. The second is named "lod" and
contains the lods for each batch.}
}
\description{
This function performs censored likelihood multiple imputation for
single-pollutant models where the pollutant of interest is subject to
varying detection limits across batches (this function will also work if
there is only one distinct detection limit). The function
outputs a list containing the imputed datasets and details regarding the
imputation procedure (i.e., number of imputed dataset, covariates used to
impute the non-detects, etc).
}
\note{
\code{clmi} only supports categorical variables that are numeric,
  (i.e., not factors or characters). You can use the  \code{model.matrix}
  function to convert data frame with factors to numeric design matrix and
  convert that matrix back into a data frame.
}
\examples{
library(lodmi)
data("toy_data")
covars <- c("smoking", "gender")
clmi.out <- clmi(toy_data, "poll", "batch1", "case_cntrl", covars,
                   lod.info = data.frame(batch.info = c("1","0"),
                     lod = c(0.8, 0.65)), 20, 12345,
                   t.function = function(x) log(x))

}
\references{
Boss J, Mukherjee B, Ferguson KK, et al. Estimating
  outcome-exposure associations when exposure biomarker detection limits
  vary across batches. To appear in \emph{Epidemiology}. 2019.
}
